fix(review): move UserSearchRelevance enum to enums directory
Address PR #850 review feedback from @MRashad26:

- packages/shared/src/enums/search.enum.ts: new home for the exported
  UserSearchRelevance enum, matching the shared enums/ convention
  (committee.enum.ts, meeting.enum.ts, ...) per CLAUDE.md (per @MRashad26)
- enums/index.ts: re-export the new enum from the barrel
- search.utils.ts / search.utils.spec.ts: import the enum from ../enums

Resolves 1 review thread.
