+++ packages/shared/src/enums/index.ts
@@ -13,3 +13,4 @@ export * from './survey.enum';
 export * from './event.enum';
 export * from './project-funding.enum';
 export * from './project-stage.enum';
+export * from './search.enum';
